requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
plotly>=5.18.0
PyQt5>=5.15.0
playsound>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.9.0
PyYAML>=6.0.0
pytest>=7.4.0
python-binance>=1.0.19
websockets>=11.0.0
SQLAlchemy>=2.0.0
loguru>=0.7.0
